refactor: don't call extension host references directly, use getReferences wrapper function instead
